Loaded into flask and SQL
To do this I created a flask folder structure along with an app.py and added in an app.route for each html page I had put inside the templates folder. I also had to change some of the file source names because of the new folder structure. Now this means flask is serving all of my pages and content correctly, and connected to my localhost database.